Add datasetIndex and chart instances to render and fontColor function object
To implement a customization on rendering and coloring you could need of information based on other elements of chart (not only dataset item). Furthermore 'index' represent the value inside the dataset but not the dataset index. Therefore I suggest to add 'datasetIndex' and 'chart' into the object to pass as argument to callbacks.